mcp_rag/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
numpy>=1.26.0
faiss-cpu>=1.7.4
